+++ NEWS.md
@@ -1,5 +1,6 @@
-# mirai 0.9.0.9000 (development)
+# mirai 0.9.0.9001 (development)
 
+* Enables secure TLS `tls+tcp://` or `wss://` connections for daemons (with automatic configuration).
 * Allows use of reference objects in a 'mirai' (such as those accessed via an external pointer), by adding argument 'refhook' to `daemons()`, `server()` and `dispatcher()` for providing a hook function to handle such objects.
 * Requires nanonext >= [0.9.0.9016].